article_server/pyproject.toml
[tool.poetry]
name = "article_server"
version = "0.1.0"
description = ""
authors = ["t4t5u0 <ymmtryk0902@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.5"
fastapi = "^0.74.1"
requests = "^2.27.1"
bs4 = "^0.0.1"
lxml = "^4.8.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
